fix Receive Siacoin UX problems
this PR fixes the UX problems with 'Receive Siacoin', namely, that each
time the dialog would open a new address would be displayed. This was
confusing to new users unfamiliar with how sia works. The behavior has
been changed so the dialog always shows the last generated address, and
includes a button to generate a new address if desired.